deepspeed-chat: filter stage3 too long prompts
In case stage3 prompts are too long, the prompts are used but they are
arbitrary sliced at start to fit into the configured max prompt length.
This arbitrary slicing sometimes causes prompts to be less meaningful.
Which in turn, causes the generator to generate garbage.
This phenomena was observed to de-stabilize RLHF stage3.
To fix it, we filter prompts that are too long.

In addition, dataset rebuild flag is propagated to other required consumers.

Change-Id: I440f09decf0784e4c2c8167a893006dff312281b
Signed-off-by: Moshe Island <misland@habana.ai>